Customer classes and add customer functionality.
A "customer" as well as "customermanager" class was created to manage the customer object.

In addition, the mainwindow class (.h, .cpp and .ui files) have been updated to accomodate these changes.

For the sake of clarity and project consistency, doxygen documentation has been added to all the new class files.