Fixed the vector handling
Before, no instance of the manager was present in the mainwindow, this has now been corrected.

I also misunderstood how the button function worked and how it interacted with the vector. These problems are now fixed and interacting with the customer objects stored in the vector should now work as intended.