UI with tableview now works
The tableview now works and shows a list with all the necessary parameters when adding customers to the database through the UI. This required quite a lot of changing the getters and setters. Debug outputs have been added as well, and although not technically necessary anymore, are left in as a precaution.

The code should serve as a decent starting point for similar functionality across the other classes.

Another key thing is that doxygen documentation was removed, as to ironically make the code more readable when working with it. It will be added back before delivery.

Filepaths were left unchanged.