Edit and remove customer works now
In addition to adding necessary functions to mainwindow.cpp to allow for these options, a new class was added, "AreYouSureDialog" which is basically a universal warning prompt to be used whenever a user intends to delete an entry, functioning as a safety feature to avoid accidental deletion.

Furthermore, the tableview was changed to allow for single row selection, which is what makes the edit and remove buttons relatively user friendly. This was as easy as checking a few boxes, just like the sorting ability.

regNr as a parameter was added to the customer class to allow for assigning a car to the customer object. It is initialized but never assigned a value, nor is it visible in any capacity to the end user.

Sidenote: I transferred the code surgically from QT, in an attempt not to mess up Blodroeds minor fixes and formatting. tried to avoid messing up filepaths and the like, as well as updating cmakelists. However, It is possible I missed something.