Car dialog plus working add car
Large commit that features a new class (cardialog) as well as updates to the pre-existing car and carmanager class to work optimally with the ui. Another parameter was also added, make, to supplement model for more presise vehicle data storage. The order of parameters was also changed slightly.

Added just for visibility is an early draft of the lease dialog window, with its accompanying leasedialog class (although no work has been put into this class as of yet, it is merely a placeholder.)